Create User IDs
A person, new to his job in a start-up, was given a task by his supervisor to prepare a
report where the user IDs (of all the employees) is a part of the report. He fetched the
user IDs from database and then used a spreadsheet function to concatenate them to
share it in email as text. However, while concatenating, he forgot to put a comma ,
in between the two user IDs
As a rule, user IDs of the portal should be minimum 4 characters and can only contain alphabets and cannot start or end with a vowel.

Supervisor knows about this rule and the number of employees in the department. Therefore, he tries to guess a possible set of user IDs from the given string (of user IDs in the email). You need to find the total no. of different set of user Ids that are possible given the number of employees in the department and the string of user IDs.