Create Earthquake Damage Detecter
Due to the latest earthquake occurrences, the government of "Intelligent Country" is trying to devise a mechanism which can accurately identify the danger zones based on the origin / epicenter of the earthquake. The mechanism requires that the local map, the co-ordinates of the epicenter and the intensity (measured in Richter scale) are provided as inputs to the program. The output of the program would be the map marked with affected areas.

The local map is represented as a grid with square dimensions containing 0's and 1's only. 0's denote the areas which are water whereas 1 denotes the land areas. Based on the epicenter co-ordinates, all the neighboring inhabited areas that would be affected needs to be marked as '2'. This process will continue until all the neighboring land areas are processed. Based on the intensity provided, the areas which would be most affected should be marked as "high risk zones" denoted by '3'. The below table should be used for reference. Intensity Impact

Table 1: Intensity-Impact Table

<=3.0 All reachable land areas would be marked as affected areas.
3.0 and <=5.0 The immediate neighbors to the epicenter are the high risk zones. 5.0 and <=8.0 The first three immediate neighbors would be marked as high risk zones. 8.0 All reachable land areas would be marked as high risk zones.