Create Mineral Samples
John is a geologist and he needs to count rock samples in order to send it to a chemical laboratory. He has a problem: The laboratory only accepts rock samples by a range of its size in ppm (parts per million).

John receives the rock samples one by one and he classifies the rock samples according to the range of the laboratory. This process is very hard because the number of rock samples may be in millions.

John needs your help, your task is to develop a program to get the number of rocks in each of the ranges accepted by the laboratory.