Create ith Largest Factor of P
A positive integer d is said to be a factor of another positive integer P if when P is divided by d, the remainder obtained is zero.

For example, for number 12, there are 6 factors 1, 2, 3, 4, 6, 12. Every positive integer i has at least two factors, 1 and the number i itself. Given two positive integers P and i, write a program to print the ith largest factor of P.