Create Prime Time Again
Here on earth, our 24-hour day is composed of two parts, each of 12hours. Each hour in each part has a corresponding hour in the other partseparated by 12 hours: the hour essentially measures the duration sincethe start of the day part. For example, 1 hour in the first part of the day is
equivalent to 13, which is 1 hour into the second part of the day.Now, consider the equivalent hours that are both prime numbers. We have 3 such instances for a 24-hour 2-part day:


5~17
7~19
11~23


Accept two natural numbers D, P >1 corresponding respectively to number
of hours per day and number of parts in a day separated by a space. D
should be divisible by P, meaning that the number of hours per part (D/P)
should be a natural number. Calculate the number of instances of
equivalent prime hours. Output zero if there is no such instance.

Note that we require each equivalent hour in each part in a day to be a prime number.

Example:

Input: 24 2
Output:3 (We have 3 instances of equivalent prime hours: 5~17, 7~19 and
11~23.)

Constraints:

10 <= D < 500
2 <= P < 50

Input:

Single line consists of two space separated integers, D and P
corresponding to number of hours per day and number of parts in a day
respectively.

Output:

Output must be a single number, corresponding to the number of
instances of equivalent prime number, as described above